backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.5.0
websockets==12.0
redis==5.0.1
openai-whisper
nltk==3.8.1
language-tool-python==2.7.1
langdetect==1.0.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.4
textblob==0.17.1